poc/b_ai_engine/poc_3a_dspy_research/requirements.txt
dspy>=3.0.4
openai>=1.58.0
python-dotenv>=1.0.1
mlflow>=2.0.0  # Observability (선택)

# System Dependencies (ProgramOfThought 사용 시)
# -------------------------------------------------
# macOS:
#   brew install deno
#
# Linux / Docker:
#   curl -fsSL https://deno.land/install.sh | sh
#   export PATH="$HOME/.deno/bin:$PATH"
